+++ fabric/src/main/java/earth/terrarium/athena/api/client/fabric/AthenaBakedModel.java
@@ -2,6 +2,7 @@ package earth.terrarium.athena.api.client.fabric;
 
 import earth.terrarium.athena.api.client.models.AthenaBlockModel;
 import earth.terrarium.athena.api.client.models.AthenaQuad;
+import earth.terrarium.athena.api.client.models.TintProvider;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import net.fabricmc.fabric.api.renderer.v1.Renderer;
 import net.fabricmc.fabric.api.renderer.v1.mesh.MutableQuadView;
@@ -55,12 +56,17 @@ public class AthenaBakedModel implements BlockStateModel, FabricBlockStateModel 
     }
 
     private void emitQuads(QuadEmitter emitter, @Nullable Direction side, List<AthenaQuad> quads) {
+        var attributes = this.model.getAttributes();
+        //noinspection deprecation
+        var layer = attributes.getLayer() != null ? attributes.getLayer() : this.model.getLayerType();
+
         for (var sprite : quads) {
             TextureAtlasSprite texture = this.textures.get(sprite.sprite());
             if (texture == null) {
                 continue;
             }
             emitter.square(side, sprite.left(), sprite.bottom(), sprite.right(), sprite.top(), sprite.depth());
+            emitter.renderLayer(layer);
 
             int flag = MutableQuadView.BAKE_LOCK_UV;
 
@@ -71,7 +77,13 @@ public class AthenaBakedModel implements BlockStateModel, FabricBlockStateModel 
             }
 
             emitter.spriteBake(texture, flag);
-            emitter.color(-1, -1, -1, -1);
+
+            switch (attributes.getTint()) {
+                case TintProvider.Index(var index) -> emitter.tintIndex(index);
+                case TintProvider.Static(var color) -> emitter.color(color, color, color, color);
+                case null -> {}
+            }
+
             emitter.emit();
         }
     }