+++ neoforge/src/main/java/earth/terrarium/athena/api/client/neoforge/AthenaModelPart.java
@@ -1,6 +1,7 @@
 package earth.terrarium.athena.api.client.neoforge;
 
 import earth.terrarium.athena.api.client.models.AthenaQuad;
+import earth.terrarium.athena.api.client.models.TintProvider;
 import earth.terrarium.athena.api.client.utils.NullableEnumMap;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import net.minecraft.client.renderer.block.model.BakedQuad;
@@ -9,6 +10,7 @@ import net.minecraft.client.renderer.chunk.ChunkSectionLayer;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.core.Direction;
 import net.minecraft.world.level.block.state.BlockState;
+import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -16,10 +18,12 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
+@ApiStatus.Internal
 public record AthenaModelPart(
         @NotNull BlockModelPart parent,
         @NotNull NullableEnumMap<Direction, Map<Direction, List<AthenaQuad>>> quads,
-        @NotNull Int2ObjectMap<TextureAtlasSprite> textures
+        @NotNull Int2ObjectMap<TextureAtlasSprite> textures,
+        @Nullable TintProvider tint
 ) implements BlockModelPart {
 
     @Override
@@ -34,7 +38,7 @@ public record AthenaModelPart(
             for (var quad : quadList) {
                 TextureAtlasSprite sprite = textures.get(quad.sprite());
                 if (sprite == null) continue;
-                quads.addAll(ForgeAthenaUtils.bakeQuad(quad, dir, sprite));
+                quads.addAll(ForgeAthenaUtils.bakeQuad(quad, dir, sprite, this.tint));
             }
         });
         return quads;