+++ neoforge/src/main/java/earth/terrarium/athena/api/client/neoforge/ForgeAthenaUtils.java
@@ -1,6 +1,7 @@
 package earth.terrarium.athena.api.client.neoforge;
 
 import earth.terrarium.athena.api.client.models.AthenaQuad;
+import earth.terrarium.athena.api.client.models.TintProvider;
 import net.minecraft.client.renderer.block.model.BakedQuad;
 import net.minecraft.client.renderer.block.model.BlockElement;
 import net.minecraft.client.renderer.block.model.BlockElementFace;
@@ -17,10 +18,10 @@ import java.util.Map;
 @ApiStatus.Internal
 public class ForgeAthenaUtils {
 
-    public static List<BakedQuad> bakeQuad(AthenaQuad quad, Direction direction, TextureAtlasSprite sprite) {
+    public static List<BakedQuad> bakeQuad(AthenaQuad quad, Direction direction, TextureAtlasSprite sprite, TintProvider tint) {
         final Vector3f start = getStartPos(quad, direction);
         final Vector3f end = getEndPos(quad, direction);
-        final BlockElementFace face = AthenaBlockElementFace.of(quad, direction, start, end);
+        final BlockElementFace face = AthenaBlockElementFace.of(quad, direction, start, end, tint);
         final BlockElement element = new BlockElement(start, end, Map.of(direction.getOpposite(), face));
         return UnbakedElementsHelper.bakeElements(
                 List.of(element),